+++ ditto/twitter/models.py
@@ -831,7 +831,7 @@ class User(TimeStampedModelMixin, DiffModelMixin, models.Model):
         try:
             return self.avatar.url
         except ValueError:
-            return static("img/default_avatar.png")
+            return static("ditto-core/img/default_avatar.png")
 
     @property
     def profile_image_url(self):