+++ tests/twitter/test_models.py
@@ -705,7 +705,7 @@ class UserTestCase(TestCase):
 
     def test_avatar_url_missing(self):
         user = UserFactory(avatar=None)
-        self.assertEqual(user.avatar_url, "/static/img/default_avatar.png")
+        self.assertEqual(user.avatar_url, "/static/ditto-core/img/default_avatar.png")
 
     def test_profile_image_url(self):
         url = "https://twitter.com/tester.jpg"