+++ CHANGES.rst
@@ -8,7 +8,7 @@ NOTE: 2020-06-20: Renamed ``master`` branch to ``main``
 Unreleased
 ----------
 
-- None
+- Fix path to `img/default_avatar.png` static file.
 
 
 1.4.0 - 2021-04-07